Use curl for HTTP requests if available
Http requests using this library currently fail if `allow_url_fopen` is
disabled. There are minor security vulnerabilities involved with leaving
that enabled so it is highly recommended to disable it.
http://phpsec.org/projects/phpsecinfo/tests/allow_url_fopen.html

Although you can take precautions to mitigate the chance of an attack,
it's safer to disable it and use curl instead.