distinguish memcache key when array of scopes are used.
Previously the memcache key would be "Google_Auth_AppIdentity::Array"
when an array of scopes was passed to AppIdentity->authenticateForScope.
This would causees the token to be overwritten when multiple
Google_Auth_AppIdentity were used with multiple scopes. The fix is to
expand the array of scopes into a string creating a unique key.