Add methods to resume Uploads
Renamed the 'getResumeUri'-Method to 'fetchResumeUri', as it's more appropriate and 'getResumeUri' is now used as a getter for the private resumeUri property. Added a 'resume'-method to be called with an resumeIri, which will fetch the upload status & setup the object. private makePutRequest method is added to avoid duplicating code from nextChunk().