Enable custom umask for cache files
While google/google-api-php-client#617 is well intentioned for its
security concerns, it has an annoying side effect when the cache is
shared between an application's CLI and Web Interface.  Unless you use
the same username for Apache or PHP-FPM to access the CLI (which you
shouldn't), a cache file will only be readable by whichever interface
created it.

My solution is to add a third constructor argument to `Google_Cache_File`
for setting a custom umask.  I've assigned it the default value `077` so
that default permissions for files and directories will remain `0600` and
`0700` respectively.