Fix listTranslations to make getTranslations work
If you make a listTranslations call you'll get a
```Google_Service_Translate_TranslationsListResponse``` back. This has a data
value which is a key with "translations" array in it. Unfortunately these aren't
returned by the getTranslations call,  you must instead use
```$translations->data['translations']```. This fixes that method so it works.